refactor: move /ecs/archetype to /types/archetype
