+++ cardinal/ecs/component/metadata/component_metadata_test.go
@@ -11,9 +11,9 @@ import (
 	"pkg.world.dev/world-engine/cardinal/ecs/filter"
 
 	"pkg.world.dev/world-engine/cardinal/ecs"
-	"pkg.world.dev/world-engine/cardinal/ecs/archetype"
 	"pkg.world.dev/world-engine/cardinal/ecs/component"
 	"pkg.world.dev/world-engine/cardinal/ecs/component/metadata"
+	"pkg.world.dev/world-engine/cardinal/types/archetype"
 
 	"pkg.world.dev/world-engine/cardinal/ecs/storage"
 )
