+++ cardinal/ecs/store/iface.go
@@ -4,12 +4,12 @@ import (
 	"encoding/json"
 	"pkg.world.dev/world-engine/cardinal/ecs/message"
 
-	"pkg.world.dev/world-engine/cardinal/ecs/archetype"
 	"pkg.world.dev/world-engine/cardinal/ecs/component/metadata"
 	"pkg.world.dev/world-engine/cardinal/ecs/entity"
 	"pkg.world.dev/world-engine/cardinal/ecs/filter"
 	ecslog "pkg.world.dev/world-engine/cardinal/ecs/log"
 	"pkg.world.dev/world-engine/cardinal/ecs/storage"
+	"pkg.world.dev/world-engine/cardinal/types/archetype"
 )
 
 type Reader interface {
